+++ code/frameworks/StreamingPipeline/lib/abstract_streaming_pipeline/GeneralStreamingPipe.py
@@ -1,18 +1,18 @@
 """
 @author           	:  rscalia                                  \n
-@build-date         :  Wed 04/08/2021                           \n
+@build-date         :  Fri 06/08/2021                           \n
 @last-update        :  Fri 06/08/2021                           \n
 
-Questa classe implementa l'astrazione della StreamingPipe presente nel DP StreamingPipeline
+Questa classe implementa una serie di metodi di base dell'interfaccia StreamingPipe.
 """
 
-from abc            import ABC,     ABCMeta, abstractmethod
-from typing         import Union ,  Iterator
+from abc                import ABC,     ABCMeta, abstractmethod
+from typing             import Union ,  Iterator
+from .StreamingPipe     import StreamingPipe
 
 
-class StreamingPipe( ABC , metaclass=ABCMeta ):
+class GeneralStreamingPipe( StreamingPipe ):
 
-    @abstractmethod
     def __init__(self, pInputStream:Iterator = None  ) -> object:
         """
         Costruttore
@@ -20,10 +20,9 @@ class StreamingPipe( ABC , metaclass=ABCMeta ):
         Args:\n
             pInputStream        (Iterator | DEF = None)        : eventuale stream di input alla Pipe
         """
-        pass
+        self._inputStream:Iterator                   = pInputStream
 
 
-    @abstractmethod
     def __iter__(self) -> Iterator:
         """
         Metodo che restituisce l'oggetto necessario al prelievo dei dati elaborati dalla pipe
@@ -31,21 +30,19 @@ class StreamingPipe( ABC , metaclass=ABCMeta ):
         Returns:\n
             Iterator                                        : Iteratore sui dati elaborati dalla Pipe
         """
-        pass
+        return self.make()
 
     
-    @abstractmethod
     def __next__(self) -> object:
         """
-        Questo metodo permette di elaborare il prossimo elemento proveniente dallo stream di input
+        Questo metodo permette di elaborare il prossimo elemento presente nello stream di input
 
         Returns:\n
-            object                                          : dato proveniente dallo stream elaborato dalla Pipe
+            object                                             : dato proveniente dallo stream elaborato dalla Pipe
         """
-        pass
+        return next ( self.make() )
 
 
-    @abstractmethod
     def hasNext(self) -> bool:
         """
         Metodo che indica se è possibile prelevare ancora elementi dallo Stream
@@ -53,10 +50,9 @@ class StreamingPipe( ABC , metaclass=ABCMeta ):
         Returns:\n
             bool                                             : VERO se è presente ancora un elemento da elaborare, FALSO altrimenti
         """
-        pass
+        return True
 
     
-    @abstractmethod
     def make(self) -> Iterator:
         """
         Restituisce un iteratore che permette di prelevare sequenzialmente i dati elaborati dalla pipe
@@ -68,10 +64,21 @@ class StreamingPipe( ABC , metaclass=ABCMeta ):
             StopIteration                               : stop dello stream verso la pipe
             Exception                                   : eccezione derivata dal prelievo dei dati dallo stream oppure dalla filter oppure dalla map oppure dall'assenza dell'oggetto per accedere allo Stream
         """
-        pass
+        while self.hasNext():
+
+            try:
+                if self._inputStream == None:
+                    raise ValueError("[!] Data Stream is unavailable !")
+
+                data:object                     = self.getFromStream()
+                if self.filter(data):
+                    computed_data:object        = self.map(data)
+                    yield computed_data
+
+            except StopIteration:
+                return
 
 
-    @abstractmethod
     def getFromStream(self) -> object:
         """
         Questo metodo preleva un elemento dallo stream di dati in input alla Pipe
@@ -82,10 +89,9 @@ class StreamingPipe( ABC , metaclass=ABCMeta ):
         Raises:\n
             Exception       : errore nella get del dato dallo stream
         """
-        pass
+        return next(self._inputStream)
 
 
-    @abstractmethod
     def __or__(self, pNextPipe:object) -> object:
         """
         Permette di connettere più Task tra di loro attraverso l'operatore Pipe (OR).
@@ -100,38 +106,5 @@ class StreamingPipe( ABC , metaclass=ABCMeta ):
             StopIteration                           : stop dello stream verso la pipe
             Exception                               : eccezione derivata dal prelievo dei dati dallo stream oppure dalla filter oppure dalla map oppure dall'assenza dell'oggetto per accedere allo Stream
         """
-        pass
-        
-        
-    @abstractmethod
-    def filter(self, pData:object) -> bool:
-        """
-        Filtra il dato attuale secondo un criterio opportuno
-        
-        Args:\n
-            pData               (object)        : dato da filtrare
-
-        Returns:\n
-            bool                                : booleano in merito alla logica del filtro oppure eccezione
-
-        Raises:\n
-            Exception                           : eventuale eccezione accaduta durante il filtraggio
-        """
-        pass
-
-
-    @abstractmethod
-    def map(self, pData:object) -> object:
-        """
-        Elabora il dato in input alla pipe
-
-        Args:\n
-            pData               (object)        : dato da elaborare
-
-        Returns:\n
-            object                              : dato elaborato o eccezzione
-
-        Raises:\n
-            Exception                           : eventuale eccezione accaduta durante l'elaborazione
-        """
-        pass
+        pNextPipe._inputStream                           = self.make()
+        return pNextPipe