+++ code/frameworks/StreamingPipeline/changelog/changelog.md
@@ -1,4 +1,8 @@
 # StreamingPipeline
 
+> **v1.0.1**
+>	* Trasformazione classe StreamingPipe in interfaccia
+>   * Creata classe astratta GeneralStreamingPipe che implementa le più comuni routine presenti nell'interfaccia StreamingPipe 
+
 > **v1.0.0**
 >	* Costruzione Componente